requirements.txt
importlib-metadata==6.5.0
langchain==0.0.136
markdown2==2.4.8
openai==0.27.2
pygments==2.15.1
pyside2~=5.15.2
python-dotenv==1.0.0
setuptools==67.6.1
